Put FSDP stuff in magic submodule
